requirements.txt
spotipy>=2.23.0
rapidfuzz>=3.0.0
unidecode>=1.3.0
rich>=13.0.0
click>=8.1.0
pyyaml>=6.0
requests>=2.31.0
deemix
